Fix cpuinfo calls, trying to fix #1896.
The compiler assumes that the output of an inline assembler block with
no input is constant. With -03 it thus starts optimizing away multiple
executions of such a block. This can be avoided by marking the
assembler block as 'volatile'.

At least for getAPIC_ID() the expected result is obviously not a
constant. It doesn't hurt though to also mark all other similar inline
assembler blocks as volatile.


git-svn-id: http://svn.ipd.uka.de/repos/ParaStation/trunk/psmgmt@8693 34291a86-62e8-0310-8159-f060f8a2ee4e