regtest.py: Add support for job arrays
This commit adds support for the job array
feature of slurm. This requires capturing the
state of different array tasks. When job arrays
are used the scontrol output for all array
tasks are exported by adding the taskid to
the prefix (e.g., task 2 in partition batch
will use the prefix PSTEST_SCONTROL_BATCH_2_).