regtest.py: Fix racy chdir
The os.chdir call lead to a race between different threads
since the current working directory is the same for all threads.
From time to time, the submit() call would fail because
sbatch/srun could not find the submission file. Further
investigation showed that the working directory had been
changed in the meantime.