regtest.py: Capture srun/sbatch exit code
So far regtest.py did not capture the exit code of the srun/sbatch
program. A comment stated that the scontrol output would contain
this information. This is howerver incorrect: The ExitCode reported
by scontrol is the exit code of the user application on the compute
node.
This commit implements this missing feature. The exit code is
provided to the evaluation process via the environment.