regetst.py: Add -r/--repetitions option
For problems of racy nature it makes sense to repeat tests multiple
times. This commit adds this capability.
Supporting this feature requires a number of changes throughout the
code. In particular we cannot just use the "output" folder in the
test directory since multiple tests can be executed concurrently in
the same folder. We introduce a test key which uniquely identifies
each execution of a test. This also simplifies correlating output
with test results since the output folder is never moved.