regtest.py: Don not raise Exception
For interactive jobs we do not expect the process to terminate
before we have figured out the jobid - After all the front end
process has not been started yet. This might happen though in
rare cases (I am not yet 100% sure when). In this case we should
not raise an Exception but try to go on.