regtest.py: Fix issue on ParTec test cluster
Due to differences in the slurm SchedParameter options the ParTec
test cluster behaves differently then Juropa-3 and does not provide
the information used in the code to extract the jobid.
This commit fixes this problem by enforcing verbose output from
the submission command. At the same time the main loop has been
reorganized to function like a state machine which simplifies the
code significantly.