regtest.py: Add safeguards
This commit adds additional safeguards to ensure that a test does
not terminate until the Slurm job is done (e.g., COMPLETED or
FAILED). This is intended to address a problem observed by ParTec
on their test cluster (Ticket #31); the root cause of the problem
has not been identified, though.