regtest.py: Take process id into account when generating the key
When running the same test multiple time in a tight loop it might
happend that the same key is generated twice. This commit fixes
this undesirable behavior by adding the process id to the string
from which the test key is generated. Different instances of
regtest.py that run in parallel will have different process ids
and thus a different key is generated.