regtest.py: Improve killability
This commit modifies the regtest.py workflow such that
SIGINT/SIGQUIT/SIGTERM are properly handled. To this end a new
member canceled of WorkerThread is introduced. If the member is
set to one threads will attempt to finish up their work as soon
as possible. In this way we can terminate the testsuite in a
clean fashion.