regtest: Allow size adaptable tests
Many tests that were written on the Juropa-3 system have hard-coded
assumptions on the number of cpus. This commit adds support for
tests that adapt the size of the submitted job based on the available
hardware. To this end, the hardware configuration (as seen by
Slurm) is queried using sinfo. srun commands can use @P on the
command line in the descr.json to reference the number of CPUs. In
sbatch jobs the environment variable PSTEST_PARTITION_CPUS is
available. The number of CPUs for each partition is also exported
for the evaluation script.
An additional new (related) feature is support for formula expansion.
Formulas encompassed by |...| in the submission line in the
description file of a test are evaluated using Python's eval()
function.