regtest.py: Change reservation/partition/qos handling
The testsuite was designed with the ability to be able to run tests
on different partitions at the same time and then compare results
in the evaluation procedure. This feature however was not very useful
in practice and had the downside that the partitions and reservations
had to be specified in the testfile. The -i/--ignorep option was
added as a workaround to allow for excluding selected partitions.

This commit introduces a backwards-incompatible change to the way
partitions and reservations are handled. At the same time support
for a qos specification is added. Instead of using the "partition"
and "reservation" specification from the test the partition, reservation
and qos can be specified on the command line using --reservation,
--partition and --qos. If --partition is not specified the default
partition is used.