Introduce psjob management
psjob is used to hold information of reservations a node is part of.
The reservation information is distributed whenever a reservation is created
and managed on the involved nodes in a psjob structure indentified by
the TID of the logger of the job. The jobs are managed in a list of local jobs
and each of them contains the list of its reservation information.