Use a designated initializer instead of a compound literal.
Designated initializers are allowed in C99, and as a GNU extension
also in C90. On the other hand, initialization of objects with static
storage duration by compound literals is a GNU extension not allowed
in C99 (because the initializer is not a constant.)

In short, this change should allow compiling with CentOS7's gcc 4.8
with and without -std=gnu99.