Rework of setting GPU binding variables
The goal of this rework was not to randomly override variables set
by the user. The problem here is, that with psslurm, we have to
support multiple start methods. Using sbatch, first the job writes
CUDA_VISIBLE_DEVICES for the case no srun is used inside the job
script. If the user then sets the variable manually inside the job
script and uses srun, too, it is impossible to descide if the user
of psslurm did set it and thus if psslurm should override it or not.
Therefor we introduce PROTECT_CUDA_VISIBLE_DEVICES for the user to
tell explicitly that he is setting the variable himself.