GPU Binding: Remove env var PROTECT_* in favour of __AUTO_*
Instead of requiring the user to explicitly set a PROTECT_*
variable to protect his manually set GPU binding variable we
now use __AUTO_* variables to detect who set the actual
variable. To handle the problem that we cannot detect an
explicit resetting of the variable to the same value, we add
some space characters which CUDA does not care about but
hopefully a user will never append by himself.