psslurm: Implement NUMA based GPU pinning
In contrast to the NUMA aware GPU pinning done by the psid
the implementation in psslurm is able to take the pinning of
sister processes into account and such use less uses GPUs and
GPUs not connected to the NUMA node the process in running on.

The heuristic is to first handle processes running on GPU
near NUMA domains. If there are multiple near GPUs to a
some processes they are bound round robin. Second each
process not running on a GPU near NUMA domain gets the
GPUs bound round robin which are bound to least other
processes. At the end each process is bound to exactly
one GPU.