psslurm: Directly generate hwthreads in partition creation
Removes the indirection over first generating a slots array
and then a hw threads array when it comes to partition generation.

This changes the threads array in two ways:
- It is not longer necessarily sorted by nodes
- It can contain the same node/thread combination twice

The first point should not be an issue since psslurm generates the
reservations, two and can assure it is matching the partition.
The second point is intended since it enables a user to explicitly
pin two processes to overlapping hwthreads by using slurm options
(--cpu-bind={map|mask}_cpu) without enabling overbooking: The slots
are generated by the pinning code from the cpu mask of map and so
there will be two slots with overlapping node/hwthreads combination
resulting in multiple times the same hw threads entry in the
partition. In the reservation those can all be used once for the
different overlapping processes.