Read real hostname from psconfig
Tries to reads <Psid.NetworkName>.Hostname from psconfig and
stores it as nodes[id].hostname in the nodes database.

If it is not found, this is not an error.

This hostname is meant to be the real hostname of the node in
the network used by the psid. As we take the IP address from
the same network, this should always be the pair the resolver
returns, too. It is a misconfiguration if not so.