psslurm: Introduce PSSLURM_VERBOSE_PINNING
If this environment variable is set, it prints additional
information to the user for debugging pinning issues.

This is meant for administrators and developers to easier
test changes especially coming with new Slurm versions.

Currently, this only prints the content of the cpuBindType
received from slurmctld, but is subject to be extended.